orgnaize character component
