+++ discordbot/components/character/start.go
@@ -1,0 +1,219 @@
+package character
+
+import (
+	"context"
+	"fmt"
+	"log"
+	"math/rand"
+	"strings"
+	"time"
+
+	"github.com/KirkDiggler/dnd-bot-go/dnderr"
+	"github.com/KirkDiggler/dnd-bot-go/internal/entities"
+	"github.com/bwmarrin/discordgo"
+	"golang.org/x/sync/errgroup"
+)
+
+func (c *Character) startNewChoices(number int) ([]*charChoice, error) {
+	// TODO cache these. cache in the client wrapper? configurable?
+	races, err := c.client.ListRaces()
+	if err != nil {
+		return nil, err
+	}
+
+	classes, err := c.client.ListClasses()
+	if err != nil {
+		return nil, err
+	}
+	log.Println("Starting new choices")
+	choices := make([]*charChoice, number)
+
+	var monk *entities.Class
+	for _, class := range classes {
+		if class.Key == "monk" {
+			monk = class
+			break
+		}
+	}
+	for idx := 0; idx < number-1; idx++ {
+		rand.Seed(time.Now().UnixNano())
+		class := classes[rand.Intn(len(classes))]
+		time.Sleep(1 * time.Nanosecond)
+		rand.Seed(time.Now().UnixNano())
+		race := races[rand.Intn(len(races))]
+		choices[idx] = &charChoice{
+			Race:  race,
+			Class: class,
+		}
+	}
+	choices[number-1] = &charChoice{
+		Race:  races[rand.Intn(len(races))],
+		Class: monk,
+	}
+
+	return choices, nil
+}
+
+// Selecting a character
+func (c *Character) handleRandomStart(s *discordgo.Session, i *discordgo.InteractionCreate) {
+	choices, err := c.startNewChoices(4)
+	if err != nil {
+		log.Println(err)
+		// TODO: Handle error
+		return
+	}
+
+	_, err = c.charManager.SaveState(context.Background(), &entities.CharacterCreation{
+		CharacterID: i.Member.User.ID,
+		LastToken:   i.Token,
+		Step:        entities.CreateStepSelect,
+	})
+	if err != nil {
+		log.Println(err)
+	}
+
+	components := make([]discordgo.SelectMenuOption, len(choices))
+	for idx, char := range choices {
+		components[idx] = discordgo.SelectMenuOption{
+			Label: fmt.Sprintf("%s the %s %s", i.Member.User.Username, char.Race.Name, char.Class.Name),
+			Value: fmt.Sprintf("choice:%d:%s:%s", idx, char.Race.Key, char.Class.Key),
+		}
+	}
+	response := &discordgo.InteractionResponse{
+		Type: discordgo.InteractionResponseChannelMessageWithSource,
+		Data: &discordgo.InteractionResponseData{
+			Flags:   discordgo.MessageFlagsEphemeral,
+			Content: "Select your new character:",
+			Components: []discordgo.MessageComponent{
+				discordgo.ActionsRow{
+					Components: []discordgo.MessageComponent{
+						discordgo.SelectMenu{
+							// Select menu, as other components, must have a customID, so we set it to this value.
+							CustomID:    selectCaracterAction,
+							Placeholder: "This is the tale of...👇",
+							Options:     components,
+						},
+					},
+				},
+			},
+		},
+	}
+	err = s.InteractionRespond(i.Interaction, response)
+	if err != nil {
+		fmt.Println(err)
+	}
+
+}
+
+func (c *Character) initializeCharacter(charID string) error {
+	log.Println("Initializing character", charID)
+	char, err := c.charManager.Get(context.Background(), charID)
+	if err != nil {
+		return err
+	}
+
+	if char.Race == nil {
+		return dnderr.NewInvalidEntityError("Race is nil")
+	}
+
+	if char.Class == nil {
+		return dnderr.NewInvalidEntityError("Class is nil")
+	}
+	g, runCtx := errgroup.WithContext(context.Background())
+
+	// Load the race starting data
+	for _, prof := range char.Race.StartingProficiencies {
+		prof := prof
+		g.Go(func() error {
+			char, err = c.charManager.AddProficiency(runCtx, char, prof)
+			if err != nil {
+				return err
+			}
+
+			return nil
+		})
+	}
+
+	for _, bonus := range char.Race.AbilityBonuses {
+		char.AddAbilityBonus(bonus)
+	}
+
+	for _, prof := range char.Class.Proficiencies {
+		prof := prof
+		g.Go(func() error {
+			char, err = c.charManager.AddProficiency(runCtx, char, prof)
+			if err != nil {
+				return err
+			}
+
+			return nil
+		})
+	}
+
+	err = g.Wait()
+	if err != nil {
+		return err
+	}
+	_, err = c.charManager.Put(context.Background(), char)
+
+	log.Println("Character initialized", charID)
+
+	return err
+}
+
+func (c *Character) handleCharSelect(s *discordgo.Session, i *discordgo.InteractionCreate) {
+	log.Println("Handling character select")
+	selectString := strings.Split(i.MessageComponentData().Values[0], ":")
+	if len(selectString) != 4 {
+		log.Printf("Invalid select string: %s", selectString)
+		return
+	}
+
+	race := selectString[2]
+	class := selectString[3]
+
+	char, err := c.charManager.Put(context.Background(), &entities.Character{
+		OwnerID: i.Member.User.ID,
+		Name:    i.Member.User.Username,
+		Race: &entities.Race{
+			Key: race,
+		},
+		Class: &entities.Class{
+			Key: class,
+		},
+	})
+	if err != nil {
+		log.Println(err)
+		return // TODO handle error
+	}
+
+	err = c.initializeCharacter(char.ID)
+	if err != nil {
+		log.Println(err)
+		return // TODO handle error
+	}
+
+	lastState, err := c.getAndUpdateState(&entities.CharacterCreation{
+		CharacterID: i.Member.User.ID,
+		LastToken:   i.Token,
+		Step:        entities.CreateStepRoll,
+	})
+	if err != nil {
+		log.Println(err)
+		return // TODO handle error
+	}
+
+	oldInteraction := &discordgo.Interaction{
+		AppID: i.AppID,
+		Token: lastState.LastToken,
+	}
+	err = s.InteractionResponseDelete(oldInteraction)
+	if err != nil {
+		log.Println(err)
+		return // TODO handle error
+	}
+
+	log.Println("Character selected", char.ID)
+
+	c.handleRollCharacter(s, i)
+}
